refactor : move seller entity in order-dataaccess to common
