+++ common/common-dataaccess/src/main/java/com/ecommerce/system/dataaccess/seller/exception/SellerDataAccessException.java
@@ -1,0 +1,7 @@
+package com.ecommerce.system.dataaccess.seller.exception;
+
+public class SellerDataAccessException extends RuntimeException {
+    public SellerDataAccessException(String message) {
+        super(message);
+    }
+}
