+++ order-service/order-dataaccess/pom.xml
@@ -20,6 +20,10 @@
             <artifactId>order-application-service</artifactId>
         </dependency>
         <dependency>
+            <groupId>com.ecommerce.system</groupId>
+            <artifactId>common-dataaccess</artifactId>
+        </dependency>
+        <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-data-jpa</artifactId>
         </dependency>
