--- order-service/order-dataaccess/src/main/java/com/ecommerce/system/order/service/dataaccess/seller/exception/SellerDataAccessException.java
@@ -1,7 +1,0 @@
-package com.ecommerce.system.order.service.dataaccess.seller.exception;
-
-public class SellerDataAccessException extends RuntimeException{
-    public SellerDataAccessException(String message) {
-        super(message);
-    }
-}
